pyproject.toml
[project]
name = "hamilton-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "ipython>=8.37.0",
    "jupyter>=1.1.1",
    "sf-hamilton>=1.88.0",
]
